Rename internal package `new_components` to `components`
