+++ src/main/kotlin/com/freya02/botcommands/api/components/ComponentGroup.kt
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.api.components
 
-import com.freya02.botcommands.internal.new_components.new.ComponentController
+import com.freya02.botcommands.internal.components.new.ComponentController
 import kotlinx.coroutines.TimeoutCancellationException
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent
