+++ src/main/kotlin/com/freya02/botcommands/api/components/EntitySelectMenu.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.api.components
 
 import com.freya02.botcommands.api.components.event.EntitySelectEvent
-import com.freya02.botcommands.internal.new_components.new.ComponentController
+import com.freya02.botcommands.internal.components.new.ComponentController
 import kotlinx.coroutines.TimeoutCancellationException
 import net.dv8tion.jda.api.interactions.components.selections.EntitySelectMenu as JDAEntitySelectMenu
