+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/AbstractComponentBuilder.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.api.components.builder
 
-import com.freya02.botcommands.internal.new_components.builder.ConstrainableComponentImpl
-import com.freya02.botcommands.internal.new_components.builder.UniqueComponentImpl
+import com.freya02.botcommands.internal.components.builder.ConstrainableComponentImpl
+import com.freya02.botcommands.internal.components.builder.UniqueComponentImpl
 
 abstract class AbstractComponentBuilder internal constructor() :
     BaseComponentBuilder,
