+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/IActionableComponent.kt
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.api.components.builder
 
-import com.freya02.botcommands.internal.new_components.ComponentHandler
+import com.freya02.botcommands.internal.components.ComponentHandler
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent
 import java.util.function.Consumer
