+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/IComponentBuilder.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.api.components.builder
 
-import com.freya02.botcommands.internal.new_components.ComponentType
-import com.freya02.botcommands.internal.new_components.LifetimeType
+import com.freya02.botcommands.internal.components.ComponentType
+import com.freya02.botcommands.internal.components.LifetimeType
 
 interface IComponentBuilder {
     val componentType: ComponentType
