+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/button/EphemeralButtonBuilder.kt
@@ -3,10 +3,10 @@ package com.freya02.botcommands.api.components.builder.button
 import com.freya02.botcommands.api.components.builder.IEphemeralActionableComponent
 import com.freya02.botcommands.api.components.builder.IEphemeralTimeoutableComponent
 import com.freya02.botcommands.api.components.event.ButtonEvent
-import com.freya02.botcommands.internal.new_components.LifetimeType
-import com.freya02.botcommands.internal.new_components.builder.EphemeralActionableComponentImpl
-import com.freya02.botcommands.internal.new_components.builder.EphemeralTimeoutableComponentImpl
-import com.freya02.botcommands.internal.new_components.new.ComponentController
+import com.freya02.botcommands.internal.components.LifetimeType
+import com.freya02.botcommands.internal.components.builder.EphemeralActionableComponentImpl
+import com.freya02.botcommands.internal.components.builder.EphemeralTimeoutableComponentImpl
+import com.freya02.botcommands.internal.components.new.ComponentController
 import net.dv8tion.jda.api.interactions.components.buttons.ButtonStyle
 
 class EphemeralButtonBuilder internal constructor(
