+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/button/PersistentButtonBuilder.kt
@@ -2,10 +2,10 @@ package com.freya02.botcommands.api.components.builder.button
 
 import com.freya02.botcommands.api.components.builder.IPersistentActionableComponent
 import com.freya02.botcommands.api.components.builder.IPersistentTimeoutableComponent
-import com.freya02.botcommands.internal.new_components.LifetimeType
-import com.freya02.botcommands.internal.new_components.builder.PersistentActionableComponentImpl
-import com.freya02.botcommands.internal.new_components.builder.PersistentTimeoutableComponentImpl
-import com.freya02.botcommands.internal.new_components.new.ComponentController
+import com.freya02.botcommands.internal.components.LifetimeType
+import com.freya02.botcommands.internal.components.builder.PersistentActionableComponentImpl
+import com.freya02.botcommands.internal.components.builder.PersistentTimeoutableComponentImpl
+import com.freya02.botcommands.internal.components.new.ComponentController
 import net.dv8tion.jda.api.interactions.components.buttons.ButtonStyle
 
 class PersistentButtonBuilder internal constructor(
