+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/group/EphemeralComponentGroupBuilder.kt
@@ -1,8 +1,8 @@
 package com.freya02.botcommands.api.components.builder.group
 
 import com.freya02.botcommands.api.components.builder.IEphemeralTimeoutableComponent
-import com.freya02.botcommands.internal.new_components.LifetimeType
-import com.freya02.botcommands.internal.new_components.builder.EphemeralTimeoutableComponentImpl
+import com.freya02.botcommands.internal.components.LifetimeType
+import com.freya02.botcommands.internal.components.builder.EphemeralTimeoutableComponentImpl
 
 class EphemeralComponentGroupBuilder internal constructor(componentIds: List<Int>) : ComponentGroupBuilder(componentIds),
     IEphemeralTimeoutableComponent by EphemeralTimeoutableComponentImpl() {
