+++ src/main/kotlin/com/freya02/botcommands/internal/components/EphemeralTimeoutHandlers.kt
@@ -1,4 +1,4 @@
-package com.freya02.botcommands.internal.new_components
+package com.freya02.botcommands.internal.components
 
 import com.freya02.botcommands.api.core.annotations.BService
