+++ src/main/kotlin/com/freya02/botcommands/internal/components/GroupTimeoutHandlers.kt
@@ -1,7 +1,7 @@
-package com.freya02.botcommands.internal.new_components
+package com.freya02.botcommands.internal.components
 
-import com.freya02.botcommands.api.components.ComponentTimeoutData
-import com.freya02.botcommands.api.components.annotations.ComponentTimeoutHandler
+import com.freya02.botcommands.api.components.GroupTimeoutData
+import com.freya02.botcommands.api.components.annotations.GroupTimeoutHandler
 import com.freya02.botcommands.api.core.annotations.BService
 import com.freya02.botcommands.internal.core.ClassPathContainer
 import com.freya02.botcommands.internal.core.requireNonStatic
@@ -10,15 +10,15 @@ import kotlin.reflect.KFunction
 import kotlin.reflect.full.findAnnotation
 
 @BService
-internal class ComponentTimeoutHandlers(classPathContainer: ClassPathContainer) : HandlerContainer {
+internal class GroupTimeoutHandlers(classPathContainer: ClassPathContainer) : HandlerContainer {
     private val map: Map<String, KFunction<*>>
 
     init {
-        map = classPathContainer.functionsWithAnnotation<ComponentTimeoutHandler>()
+        map = classPathContainer.functionsWithAnnotation<GroupTimeoutHandler>()
             .requireNonStatic()
-            .withFirstArg(ComponentTimeoutData::class)
+            .withFirstArg(GroupTimeoutData::class)
             .associate {
-                it.function.findAnnotation<ComponentTimeoutHandler>()!!.name to it.function
+                it.function.findAnnotation<GroupTimeoutHandler>()!!.name to it.function
             }
     }
