+++ src/main/kotlin/com/freya02/botcommands/internal/components/LifetimeType.kt
@@ -1,4 +1,4 @@
-package com.freya02.botcommands.internal.new_components
+package com.freya02.botcommands.internal.components
 
 import com.freya02.botcommands.internal.throwUser
