+++ src/main/kotlin/com/freya02/botcommands/internal/components/builder/EphemeralActionableComponentImpl.kt
@@ -1,7 +1,7 @@
-package com.freya02.botcommands.internal.new_components.builder
+package com.freya02.botcommands.internal.components.builder
 
 import com.freya02.botcommands.api.components.builder.IEphemeralActionableComponent
-import com.freya02.botcommands.internal.new_components.EphemeralHandler
+import com.freya02.botcommands.internal.components.EphemeralHandler
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent
 
 internal class EphemeralActionableComponentImpl<E : GenericComponentInteractionCreateEvent> : IEphemeralActionableComponent<E> {
