+++ src/main/kotlin/com/freya02/botcommands/internal/components/builder/EphemeralTimeoutableComponentImpl.kt
@@ -1,7 +1,7 @@
-package com.freya02.botcommands.internal.new_components.builder
+package com.freya02.botcommands.internal.components.builder
 
 import com.freya02.botcommands.api.components.builder.IEphemeralTimeoutableComponent
-import com.freya02.botcommands.internal.new_components.new.EphemeralTimeout
+import com.freya02.botcommands.internal.components.new.EphemeralTimeout
 import kotlinx.datetime.Clock
 import kotlin.time.Duration
