+++ src/main/kotlin/com/freya02/botcommands/internal/components/new/ComponentData.kt
@@ -1,9 +1,9 @@
-package com.freya02.botcommands.internal.new_components.new
+package com.freya02.botcommands.internal.components.new
 
 import com.freya02.botcommands.api.components.InteractionConstraints
-import com.freya02.botcommands.internal.new_components.ComponentHandler
-import com.freya02.botcommands.internal.new_components.ComponentType
-import com.freya02.botcommands.internal.new_components.LifetimeType
+import com.freya02.botcommands.internal.components.ComponentHandler
+import com.freya02.botcommands.internal.components.ComponentType
+import com.freya02.botcommands.internal.components.LifetimeType
 
 internal abstract class ComponentData(
     val componentId: Int,
