+++ src/main/kotlin/com/freya02/botcommands/internal/components/new/PersistentComponentData.kt
@@ -1,9 +1,9 @@
-package com.freya02.botcommands.internal.new_components.new
+package com.freya02.botcommands.internal.components.new
 
 import com.freya02.botcommands.api.components.InteractionConstraints
-import com.freya02.botcommands.internal.new_components.ComponentType
-import com.freya02.botcommands.internal.new_components.LifetimeType
-import com.freya02.botcommands.internal.new_components.PersistentHandler
+import com.freya02.botcommands.internal.components.ComponentType
+import com.freya02.botcommands.internal.components.LifetimeType
+import com.freya02.botcommands.internal.components.PersistentHandler
 
 internal class PersistentComponentData(
     componentId: Int,
