+++ src/main/kotlin/com/freya02/botcommands/internal/components/new/repositories/ComponentRepository.kt
@@ -1,4 +1,4 @@
-package com.freya02.botcommands.internal.new_components.new.repositories
+package com.freya02.botcommands.internal.components.new.repositories
 
 import com.freya02.botcommands.api.components.Components
 import com.freya02.botcommands.api.components.InteractionConstraints
@@ -6,10 +6,10 @@ import com.freya02.botcommands.api.components.builder.BaseComponentBuilder
 import com.freya02.botcommands.api.components.builder.ITimeoutableComponent
 import com.freya02.botcommands.api.components.builder.group.ComponentGroupBuilder
 import com.freya02.botcommands.api.core.annotations.ConditionalService
+import com.freya02.botcommands.internal.components.*
+import com.freya02.botcommands.internal.components.new.*
 import com.freya02.botcommands.internal.core.db.Database
 import com.freya02.botcommands.internal.core.db.Transaction
-import com.freya02.botcommands.internal.new_components.*
-import com.freya02.botcommands.internal.new_components.new.*
 import com.freya02.botcommands.internal.rethrowUser
 import com.freya02.botcommands.internal.throwInternal
 import com.freya02.botcommands.internal.throwUser
